[Feat] #61 - 탐색 탭 전환 구현
