+++ Napzak-iOS/Napzak-iOS/Application/RootView.swift
@@ -10,12 +10,14 @@ import SwiftUI
 struct RootView: View {
     @StateObject private var authRouter = AuthNavigationRouter()
     @StateObject private var navigationRouter = NavigationRouter()
+    @StateObject private var tabRouter = TabRouter()
     @StateObject private var authManager = AuthManager.shared
     
     var body: some View {
         if authManager.isAuthenticated && !authManager.needsOnboarding {
             NZTabBarView()
                 .environmentObject(navigationRouter)
+                .environmentObject(tabRouter)
         } else {
             LoginView()
                 .environmentObject(authRouter)
