+++ Napzak-iOS/Napzak-iOS/Global/Enums/NZTab.swift
@@ -1,0 +1,15 @@
+//
+//  NZTab.swift
+//  Napzak-iOS
+//
+//  Created by 조호근 on 5/11/25.
+//
+
+import Foundation
+
+enum NZTab {
+    case home
+    case search
+    case chat
+    case my
+}
