+++ Napzak-iOS/Napzak-iOS/Presentation/TabBar/View/NZTabBarView.swift
@@ -8,21 +8,11 @@
 import SwiftUI
 
 struct NZTabBarView: View {
-    
-    //MARK: - Enum
-    
-    enum NZTab {
-        case home
-        case search
-        case chat
-        case my
-    }
-    
     //MARK: - Property Wrappers
     
     @EnvironmentObject private var navigationRouter: NavigationRouter
+    @EnvironmentObject private var tabRouter: TabRouter
     
-    @State private var selectedTab: NZTab = .home
     @State private var isRegisterTabSelected = false
     @State private var isRegisterViewPresented = false
     @State private var registerType: TradeType = .sell
@@ -34,7 +24,7 @@ struct NZTabBarView: View {
     var body: some View {
         NavigationStack(path: $navigationRouter.path) {
             ZStack(alignment: .bottom) {
-                TabView(selection: $selectedTab) {
+                TabView(selection: $tabRouter.selectedTab) {
                     Group {
                         HomeView()
                             .tag(NZTab.home)
@@ -106,7 +96,7 @@ struct NZTabBarView: View {
     var tabBar: some View {
         HStack {
             Button {
-                selectedTab = .home
+                tabRouter.switchToHome()
                 isRegisterTabSelected = false
             } label: {
                 VStack(alignment: .center, spacing: 5) {
@@ -121,7 +111,7 @@ struct NZTabBarView: View {
             }
             Spacer()
             Button {
-                selectedTab = .search
+                tabRouter.switchToSearch()
                 isRegisterTabSelected = false
             } label: {
                 VStack(alignment: .center, spacing: 5) {
@@ -150,7 +140,7 @@ struct NZTabBarView: View {
             }
             Spacer()
             Button {
-                selectedTab = .chat
+                tabRouter.switchToChat()
                 isRegisterTabSelected = false
             } label: {
                 VStack(alignment: .center, spacing: 5) {
@@ -165,7 +155,7 @@ struct NZTabBarView: View {
             }
             Spacer()
             Button {
-                selectedTab = .my
+                tabRouter.switchToMy()
                 isRegisterTabSelected = false
             } label: {
                 VStack(alignment: .center, spacing: 5) {
@@ -195,7 +185,7 @@ private extension NZTabBarView {
     //MARK: - Private Method
     
     func isSelectedTab(_ tab: NZTab) -> Bool {
-        return selectedTab == tab && !isRegisterTabSelected
+        return tabRouter.selectedTab == tab && !isRegisterTabSelected
     }
 }
